[declarative] Reverting mapping change erroneously introduced in r9168 for domain.Report4Sitting, this fixes the observed error obtained when additing a sitting.
Removing the "group_id" field from base DocumentDescriptor, moving it down to let selected custom doc types (Question, Bill and Event) define it, as this is (a) superfluous for the other types, and (b) setting it for them will interfere with the proper function of how sscheduling filters on this filed to select docs for scheduling; also seems to fix the observed error when adding an AgendaItem. 
Updating bungeni_custom ui.xml.
Adding a revisit comment concerning the conceptual roles/interaction of parliament_id and group_id.

Switching BungeniAttributeDisplay template for easier debugging of the observed error when displaying a Ministry (for easier debugging in general, including debugging the other errors above) and to turn the page-crashing error to be limit it only to the sub-content concerned, so that a ministry may still be viewed (and transited to active, and be used by questions, etc), for tests to be able to proceed further.

Some src tweaking in ui/calendar/data.py.