[workflow] Reworking Address workflow to add an initial draft state into which all newly created addresses are transited to (when parent is NOT in draft), and dropping all global grants (as per guideline) (to be consistent with same pattern used for Attachment in r9306 and Event in r9273).
Updating workflows for head types that support Address i.e. User and Group, to manage Address sub-item permissions (as per guidelines).

Updating auto-generated workflow permissions.

Correcting (and commenting) an issue in the display logic for submenus, that was causing menus with "empty" submenus to not be displayed (issue introduced in r9327).