[schema change] Adding User.marital_status emuneration field; adding same-named vdex vocabulary.
Changing the definition of the derived User.fullname property; 

Making it possible for vdex vocabaluary *nullable* fields to be translated in a listing; changing listing column renderer descriptor.vocabulary_column() to handle None term of vocabulary, for when (a) the field allows a None value and (b) for when a None entry is not defined by the (vdex) vocabulary).
Adding NONE_LOOKUPERROR comments, as this issue needs further work.

Changing definition of User.fullname property to also include middle_name; switching two instances of redefinitions of this in ui.rest to use the property defined on User, and some src restyling.