[timestamp] Deleting timestamp as a UI concept.
A db timestamp is not a UI concern, and the now simplified and cleaner UI code makes it easy to see that that UI code concerning timestamps are just kludges to workaround some previously real or perceived  issue (for which no pertinent src comments are available).

Removing this code corrects the error that is occurring--since r9954 (that reworks how permissions are set per UI attribute)--when the minister submits a question response, that was throwing an aunauthorized error on attempt to set the "timesamp" attribute.